autodialer_bot/requirements.txt
python-telegram-bot==20.7
python-telegram-bot[job-queue]==20.7
paramiko==3.4.0
gTTS==2.4.0
pydub==0.25.1
asyncio==3.4.3
aiofiles==23.2.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1
pyst2==0.5.1
asterisk-ami==0.1.5
panoramisk==1.4
cryptography==41.0.7
